Include persistent depressive disorder in caseness criteria
We have taken the reviewer’s advice and included persistent depressive disorder (SNOMED-CT: 1153575004) in the set of caseness inclusion diagnoses. We apologise for our oversight in not including it as an important sibling code in the SNOMED-CT hierarchy. We first applied the suggest change to the exclusion codelist, and have presented the ramifications elsewhere in our response.
We were surprised to note that the no records in our dataset contained a code for persistent depressive disorder. We extended the search to all records within Connected Bradford and, again, found no record containing a code for persistent depressive disorder. We confirmed that persistent depressive disorder is not on the list of excluded codes, so we can only conclude that this single, childless code is not used by general practitioners in the Bradford and Airedale region.