Add sensitivity analysis for `Rx_window_caseness`
A peer reviewer asked us to check how sensitive the count of records satisfying caseness was to he duration of the search window used to detect recent prescriptions of interest.

My conclusions are that the different (calculated on 19th March 2024) is, on average 56 records per extra year in duration, from 1 year to 10 years.