Update figure 2 caption
A peer reviewer noted that the names of the features being referred to in the figure were not included in the plot. This was because they were named in the figure caption that was supplied to the publisher.
This commit updates the figures to include the names of the feature to which the plots refer.